Improved ProfileLikelihoodResult
Now the admissible range for xi and the return level are taken into account. As a result, the getParameterConfidenceInterval() method throws if the bounds are not within the admissible range.

The drawProfileLikelihoodFunction() method has been improved by reducing the graph to its portion where the curve is at most at the same distance of the threshold as its maximum is. This way, if the function goes deeper, the graph will be symmetrical wrt the threshold.