Improved GeneralizedExtremeValueFactory
Now, the time varying estimation can be initialized using either the static estimator or the pseudo-Gumbel estimator proposed in the ismev R package. It is also possible to add a normalization step, using an affine transformation mapping the time to [0, 1] or such that the mean of t is zero and the standard deviation of t is 1, or no transformation.